+++ src/main/java/az/hrportal/hrportalapi/dto/document/EmployeeDocumentInformation.java
@@ -5,8 +5,6 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @FieldDefaults(level = AccessLevel.PRIVATE)
@@ -14,9 +12,9 @@ public class EmployeeDocumentInformation {
     String departmentName;
     String subDepartmentName;
     String vacancyName;
-    BigDecimal salary;
-    BigDecimal additionalSalary;
-    BigDecimal ownAdditionalSalary;
+    float salary;
+    float additionalSalary;
+    float ownAdditionalSalary;
     String fullName;
     String workMode;
 }
