+++ src/main/java/az/hrportal/hrportalapi/dto/employee/response/EmployeeResponseDto.java
@@ -5,8 +5,6 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @FieldDefaults(level = AccessLevel.PRIVATE)
@@ -17,5 +15,5 @@ public class EmployeeResponseDto {
     String department;
     String position;
     Boolean active;
-    BigDecimal salary;
+    float salary;
 }
